feat: profile scaffold, remove badges from profile
